average-sessions/requirements.txt
google-cloud-pubsub
google-cloud-bigtable
google-cloud